python_backend/backend/requirements.txt
asgiref==3.4.1
Django==3.2.15
django-cors-headers==3.10.1
djangorestframework==3.13.1
gunicorn==20.1.0
numpy==1.20.3
pandas==1.1.5
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
djangorestframework-simplejwt==5.2.0
typing-extensions==4.1.1
